gpu-processor/requirements.txt
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.24.3
pillow==10.2.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
